requirements.txt
pandas
numpy
matplotlib
scikit-learn
tensorflow
streamlit
pillow
